Changes reservations report to return num of reservations

+++ backend/src/main/java/team10/app/controller/BusinessClientController.java
@@ -85,7 +85,7 @@ public class BusinessClientController {
 
     @GetMapping("/report/reservations")
     @PreAuthorize("hasAnyRole('HOUSE_OWNER', 'SHIP_OWNER', 'FISHING_INSTRUCTOR')")
-    public ResponseEntity<Map<String, Double>> getEarningsReport(@RequestHeader(name = "Authorization") String token,
+    public ResponseEntity<Map<String, Integer>> getEarningsReport(@RequestHeader(name = "Authorization") String token,
                                                                  @RequestParam String period)
     {
         try {
